Fix hard-coded override of interactive flag
The 'interactive' parameter got overwritten in
polyparticletracker_parallelx, causing a figure always to be created.
This was particlarly bad for running compiled code on the CSF, because
the job would never finish with the figure still open.

This commit fixes this bug, so the plotting is switched on and off
by the argument supplied when the function is called.